refactor ChainAmountIndexer
